Updating `transposition` file to `expNN_transposition
