--- python/diagonal_elements.py
@@ -1,5 +1,0 @@
-
-
-
-
-
